refactor: timer
